+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 - ZLIB support via new `ZlibCompressor` and `ZlibDecompressor` provided by `kompress-zlib`
 - `zlibSink` and `zlibSource` extensions for `ZlibCompressor`
 - `unzlibSink` and `unzlibSource` extensions for `ZlibDecompressor`
-- `WrappingCompressor` delegate compressor to support wrapping existing compressors with extra data
-- `UnwrappingDecompressor` delegate decompressor to support unwrapping extra data with existing decompressors
+- `FramingCompressor` delegate compressor to support wrapping existing compressors with extra data
+- `FramingDecompressor` delegate decompressor to support unwrapping extra data with existing decompressors
 - Documentation for `kompress-gzip` APIs
 - Documentation for `kompress-zip` APIs
 - Documentation for `kompress-zlib` APIs